Revert "2025/07/01 요약페이지 구현"
